Updated SDK from UE4.27 to UE5.1 (#10)
* Updated SDK from UE4.27 to UE5.1

Updated the project from 4.27 to 5.1. Added .vsconfig to the .gitignore. Updated/Added some some of the necessary .ini files.

* Updated Log category across native code

Updated the LogCategory for UE_LOG's and error handling to LogPlayerWON so that it can be filtered via the UE5 console/output.

Switched a few print to screen functions out with Logs as they're less expensive and easier to catch.