Sdk update (#11)
* Updated SDK from UE4.27 to UE5.1

Updated the project from 4.27 to 5.1. Added .vsconfig to the .gitignore. Updated/Added some some of the necessary .ini files.

* Updated Log category across native code

Updated the LogCategory for UE_LOG's and error handling to LogPlayerWON so that it can be filtered via the UE5 console/output.

Switched a few print to screen functions out with Logs as they're less expensive and easier to catch.

* Added new fields to FClientDetails and FSessionDetails

Added the needed properties to the FClientDetails and FSessionDetails structs and updated the BlueprintLibrary functions with the new properties.

* Debugged and tested to make sure API still works.  Added default test ad.

Debugged and tested the API using credentials provided by Adam. Added in the default ad and made both example levels use the default ad directing people to contact PlayerWON for paid ads.