Updating plugin to ue5.5 (#13)
* Removed old plugin and replaced with new

Removed the old plugin made with UE4.27 and replaced it with migrated code and assets to UE5.5.  Included plugin builders with comments.

* Delete .vsconfig as it's not needed in the repo

Deleting the .vsconfig from the repo. It's not needed and should be ignored.

* Updated main gitignore

updated main gitignore to look for the new Unreal plugin.

* Updated readme and build file

Updated the readme and the build 5.5 file.

* Updated uplugin with needed info.

Updated the DocsURL with the readme